# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cumulocity Service Request API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "basicAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/context/config/{configId}" : {
      "get" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "GET context configuration by Id",
        "description" : "Returns context configuration by internal Id",
        "operationId" : "getContextConfigById",
        "parameters" : [ {
          "name" : "configId",
          "in" : "path",
          "description" : "Internal context configuration Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "UPDATE context configuration",
        "description" : "Updates an existing context configuration.",
        "operationId" : "updateContextConfig",
        "parameters" : [ {
          "name" : "configId",
          "in" : "path",
          "description" : "Internal context configuration Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContextConfig"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "DELETE context configuration by Id",
        "description" : "Deletes a context configuration by Id",
        "operationId" : "deleteContextConfigById",
        "parameters" : [ {
          "name" : "configId",
          "in" : "path",
          "description" : "Internal context configuration Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/context/config" : {
      "get" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "GET all context configurations",
        "description" : "Returns complete list of context configurations",
        "operationId" : "getContextConfigList",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ContextConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "CREATE context configuration",
        "description" : "Creates a new context configuration for alarm data context.",
        "operationId" : "createContextConfig",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContextConfig"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextConfig"
                }
              }
            }
          }
        }
      }
    },
    "/api/context/config/apply/alarm/{alarmId}" : {
      "post" : {
        "tags" : [ "context-config-controller" ],
        "summary" : "APPLY context configurations to alarm",
        "description" : "Applies all matching context configurations to a specific alarm by alarm ID.",
        "operationId" : "applyContextConfigsToAlarm",
        "parameters" : [ {
          "name" : "alarmId",
          "in" : "path",
          "description" : "Internal alarm Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Not Found - Alarm not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error - Context configuration is not valid",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "200" : {
            "description" : "OK - Context configurations applied successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ContextApplyRules" : {
        "type" : "object",
        "description" : "Rules that define when this context configuration should be applied to an alarm",
        "properties" : {
          "devicePredicate" : {
            "type" : "array",
            "description" : "Device predicates that must match for this context to be applied",
            "items" : {
              "$ref" : "#/components/schemas/ContextPredicate"
            }
          },
          "alarmPredicate" : {
            "type" : "array",
            "description" : "Alarm predicates that must match for this context to be applied",
            "items" : {
              "$ref" : "#/components/schemas/ContextPredicate"
            }
          }
        }
      },
      "ContextConfig" : {
        "type" : "object",
        "description" : "Context Configuration object that defines rules and settings for creating contextualized alarms. Contains device and alarm predicates to determine when to apply the configuration.",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Internal ID, set by Cumulocity",
            "example" : 123456789,
            "readOnly" : true
          },
          "description" : {
            "type" : "string",
            "description" : "Description of what this context configuration is used for",
            "example" : "Context configuration for CS82 control software connection errors"
          },
          "apply" : {
            "$ref" : "#/components/schemas/ContextApplyRules",
            "description" : "Application rules that define when this context configuration should be applied"
          },
          "config" : {
            "$ref" : "#/components/schemas/ContextSettings",
            "description" : "Context configuration settings including time windows and data selections"
          },
          "isActive" : {
            "type" : "boolean",
            "description" : "Whether this context configuration is active and should be used",
            "example" : true
          }
        },
        "required" : [ "apply", "config", "description", "id" ]
      },
      "ContextPredicate" : {
        "type" : "object",
        "description" : "Predicate for matching objects based on fragment values",
        "properties" : {
          "fragment" : {
            "type" : "string",
            "description" : "Fragment name to check",
            "example" : "ec_ControlSoftware"
          },
          "regex" : {
            "type" : "string",
            "description" : "Regular expression that the fragment value must match",
            "example" : "^CS82$"
          }
        },
        "required" : [ "fragment", "regex" ]
      },
      "ContextSettings" : {
        "type" : "object",
        "description" : "Context configuration settings that define the data and time window",
        "properties" : {
          "dateFrom" : {
            "type" : "string",
            "description" : "Relative start time for the context window",
            "example" : "-1d"
          },
          "dateTo" : {
            "type" : "string",
            "description" : "Relative end time for the context window",
            "example" : "0d"
          },
          "datapoints" : {
            "type" : "array",
            "description" : "List of measurement data points to include in the context",
            "example" : [ "MHAI1.H1RMSA1", "MHAI1.H1RMSA2" ],
            "items" : {
              "type" : "string"
            }
          },
          "events" : {
            "type" : "array",
            "description" : "List of event types to include in the context",
            "example" : [ "ec_S:242/232", "ec_S:66/11" ],
            "items" : {
              "type" : "string"
            }
          },
          "alarms" : {
            "type" : "array",
            "description" : "List of alarm types to include in the context",
            "example" : [ "ec_S:242/232" ],
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "dateFrom", "dateTo" ]
      },
...
      "ErrorResponseBody" : {
        "type" : "object",
        "description" : "Response body for error. This can contain additional error message for the client.",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          }
        }
      },
...
      "ContextData" : {
        "type" : "object",
        "description" : "Context data containing time-based contextual information with measurements, events, and alarms for a specific time window",
        "properties" : {
          "dateFrom" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Start date and time of the context window",
            "example" : "2025-09-20T18:12:09.000"
          },
          "dateTo" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "End date and time of the context window",
            "example" : "2025-09-21T18:12:09.000"
          },
          "datapoints" : {
            "type" : "array",
            "description" : "List of measurement datapoints included in this context",
            "example" : [ "MHAI1.H1RMSA1", "MHAI1.H1RMSA2" ],
            "items" : {
              "type" : "string"
            }
          },
          "events" : {
            "type" : "array",
            "description" : "List of event types included in this context",
            "example" : [ "ec_S:66/11", "ec_S:66/11" ],
            "items" : {
              "type" : "string"
            }
          },
          "alarms" : {
            "type" : "array",
            "description" : "List of alarm types included in this context",
            "items" : {
              "type" : "string"
            }
          },
          "contextConfigId" : {
            "type" : "string",
            "description" : "ID of the context configuration that generated this context data",
            "example" : 123456789
          }
        },
        "required" : [ "contextConfigId", "dateFrom", "dateTo" ]
      },
...
    },
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}